front: move train selection logic from Timetable to TimetableBoardWrapper
