+++ front/public/locales/en/operational-studies.json
@@ -233,6 +233,7 @@
       "resetAllExceptions": "Do you confirm the reset of the exceptions of this service?",
       "resetExceptions": "Reset exceptions",
       "rollingStockFilterPlaceholder": "Family, detail, serie",
+      "selectAll": "Select all",
       "showAllCategories": "All",
       "showAllTrains": "All",
       "showHonoredTrains": "Honored",
@@ -249,6 +250,7 @@
       "trainDeleted": "{{ name }} train has been deleted.",
       "trainSchedule": "Unique train",
       "trainType": "Services, trains",
+      "unselectAll": "Unselect all",
       "update": "Edit",
       "validityFilter": "Trains validity"
     },
