+++ front/public/locales/fr/operational-studies.json
@@ -233,6 +233,7 @@
       "resetAllExceptions": "Confirmez-vous la réinitialisation des exceptions de cette mission?",
       "resetExceptions": "Réinitialiser les exceptions",
       "rollingStockFilterPlaceholder": "Famille, détail, série",
+      "selectAll": "Tout sélectionner",
       "showAllCategories": "Tous",
       "showAllTrains": "Tous",
       "showHonoredTrains": "Honorée",
@@ -249,6 +250,7 @@
       "trainDeleted": "Le train {{name}} a bien été supprimé.",
       "trainSchedule": "Train unique",
       "trainType": "Missions, trains",
+      "unselectAll": "Tout désélectionner",
       "update": "Modifier",
       "validityFilter": "Validité des trains"
     },
