+++ front/src/styles/scss/applications/operationalStudies/_scenario.scss
@@ -530,18 +530,6 @@
               padding-bottom: 12px;
             }
 
-            .train-count {
-              display: flex;
-              align-items: center;
-              flex-grow: 1;
-              height: 24px;
-
-              .label {
-                margin-left: 9px;
-                color: var(--grey80);
-                font-size: 1rem;
-              }
-            }
             .more-details-button {
               font-size: 0.875rem;
               color: var(--primary60);
